+++ bot/src/main/kotlin/tech/pucci/bot/DiscordBot.kt
@@ -1,0 +1,6 @@
+package tech.pucci.bot
+
+interface DiscordBot {
+
+    suspend fun initialize()
+}
