+++ bot/src/main/kotlin/tech/pucci/usecase/GetBotAuthTokenUseCase.kt
@@ -1,0 +1,6 @@
+package tech.pucci.usecase
+
+interface GetBotAuthTokenUseCase {
+
+    val token: String
+}
