+++ command/src/main/kotlin/tech/pucci/Command.kt
@@ -1,0 +1,10 @@
+package tech.pucci
+
+import dev.kord.core.entity.Message
+
+interface Command {
+
+    val commandMap: Pair<String, Command>
+
+    suspend fun execute(message: Message)
+}
