+++ command/src/main/kotlin/tech/pucci/ping/PingCommand.kt
@@ -1,0 +1,19 @@
+package command.ping
+
+import tech.pucci.Command
+import dev.kord.core.entity.Message
+
+class PingCommand : Command {
+
+    companion object {
+        private const val COMMAND = "!ping"
+        private const val RESPONSE = "pong"
+    }
+
+    override val commandMap: Pair<String, Command>
+        get() = COMMAND to this
+
+    override suspend fun execute(message: Message) {
+        if (message.content == COMMAND) message.channel.createMessage(RESPONSE)
+    }
+}
