+++ command/src/main/kotlin/tech/pucci/waifu/api/AnimeGirlsWithBooksApi.kt
@@ -1,0 +1,8 @@
+package tech.pucci.waifu.api
+
+import io.ktor.client.statement.HttpResponse
+
+interface AnimeGirlsWithBooksApi {
+
+    suspend fun getRandomWaifu(): HttpResponse
+}
