+++ command/src/main/kotlin/tech/pucci/waifu/model/AnimeGirlsWithBookResponse.kt
@@ -1,0 +1,8 @@
+package tech.pucci.waifu.model
+
+import io.ktor.utils.io.ByteReadChannel
+
+data class AnimeGirlsWithBookResponse(
+    val fileExtension: String,
+    val image: ByteReadChannel
+)
