+++ command/src/main/kotlin/tech/pucci/waifu/repository/AnimeGirlsWithBooksRepository.kt
@@ -1,0 +1,9 @@
+package tech.pucci.waifu.repository
+
+import kotlinx.coroutines.flow.Flow
+import tech.pucci.waifu.model.AnimeGirlsWithBookResponse
+
+interface AnimeGirlsWithBooksRepository {
+
+    suspend fun fetchRandomWaifu(): Flow<Result<AnimeGirlsWithBookResponse>>
+}
