+++ eventhandler/src/main/kotlin/tech/pucci/EventHandler.kt
@@ -1,0 +1,8 @@
+package tech.pucci
+
+import dev.kord.core.event.Event
+
+interface EventHandler<T : Event> {
+
+    suspend fun handle(event: T)
+}
