+++ eventhandler/src/main/kotlin/tech/pucci/module/eventHandlerModule.kt
@@ -1,0 +1,8 @@
+package tech.pucci.module
+
+import tech.pucci.MessageCreateEventHandler
+import org.koin.dsl.module
+
+val eventHandlerModule = module {
+    factory { MessageCreateEventHandler(pingCommand = get(), waifuCommand = get()) }
+}
