+++ network/src/main/kotlin/tech/pucci/exception/FileExtensionNotProvidedException.kt
@@ -1,0 +1,11 @@
+package tech.pucci.exception
+
+class FileExtensionNotProvidedException : NetworkErrorException(Exception()) {
+
+    companion object {
+        private const val MESSAGE = "File extensions was not provided on the content header"
+    }
+
+    override val message: String
+        get() = MESSAGE
+}
