+++ settings.gradle.kts
@@ -1,3 +1,6 @@
 
 rootProject.name = "xinxila-bot"
-
+include("network")
+include("bot")
+include("eventhandler")
+include("command")
