--- src/main/kotlin/command/Command.kt
@@ -1,10 +1,0 @@
-package command
-
-import dev.kord.core.entity.Message
-
-interface Command {
-
-    val commandMap: Pair<String, Command>
-
-    suspend fun execute(message: Message)
-}
