--- src/main/kotlin/command/waifu/api/AnimeGirlsWithBooksApi.kt
@@ -1,8 +1,0 @@
-package command.waifu.api
-
-import io.ktor.client.statement.HttpResponse
-
-interface AnimeGirlsWithBooksApi {
-
-    suspend fun getRandomWaifu(): HttpResponse
-}
