--- src/main/kotlin/command/waifu/model/AnimeGirlsWithBookResponse.kt
@@ -1,8 +1,0 @@
-package command.waifu.model
-
-import io.ktor.utils.io.ByteReadChannel
-
-data class AnimeGirlsWithBookResponse(
-    val fileExtension: String,
-    val image: ByteReadChannel
-)
