--- src/main/kotlin/command/waifu/repository/AnimeGirlsWithBooksRepository.kt
@@ -1,9 +1,0 @@
-package command.waifu.repository
-
-import command.waifu.model.AnimeGirlsWithBookResponse
-import kotlinx.coroutines.flow.Flow
-
-interface AnimeGirlsWithBooksRepository {
-
-    suspend fun fetchRandomWaifu(): Flow<Result<AnimeGirlsWithBookResponse>>
-}
