--- src/main/kotlin/discord/bot/DiscordBot.kt
@@ -1,6 +1,0 @@
-package discord.bot
-
-interface DiscordBot {
-
-    suspend fun initialize()
-}
