--- src/main/kotlin/discord/usecase/GetBotAuthTokenUseCase.kt
@@ -1,6 +1,0 @@
-package discord.usecase
-
-interface GetBotAuthTokenUseCase {
-
-    val token: String
-}
