--- src/main/kotlin/eventhandler/EventHandler.kt
@@ -1,8 +1,0 @@
-package eventhandler
-
-import dev.kord.core.event.Event
-
-interface EventHandler<T: Event> {
-
-    suspend fun handle(event: T)
-}
