--- src/main/kotlin/eventhandler/module/eventHandlerModule.kt
@@ -1,8 +1,0 @@
-package eventhandler.module
-
-import eventhandler.MessageCreateEventHandler
-import org.koin.dsl.module
-
-val eventHandlerModule = module {
-    factory { MessageCreateEventHandler(pingCommand = get(), waifuCommand = get()) }
-}
