--- src/main/kotlin/main.kt
@@ -1,24 +1,0 @@
-import command.module.commandModule
-import discord.bot.DiscordBot
-import discord.module.discordModule
-import eventhandler.module.eventHandlerModule
-import network.module.networkModule
-import org.koin.core.context.GlobalContext
-import org.koin.core.context.startKoin
-
-suspend fun main() {
-
-    startKoin {
-        modules(
-            listOf(
-                networkModule,
-                commandModule,
-                eventHandlerModule,
-                discordModule
-            )
-        )
-    }
-
-    val discordBot: DiscordBot = GlobalContext.get().get()
-    discordBot.initialize()
-}
