--- src/main/kotlin/network/module/networkModule.kt
@@ -1,9 +1,0 @@
-package network.module
-
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.cio.CIO
-import org.koin.dsl.module
-
-val networkModule = module {
-    factory { HttpClient(CIO) }
-}
